fixed tests for delegate utils
